# apps/api-gateway/src/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Performa Studio API
  description: API documentation for the Performa Studio platform
  version: 1.0.0
  contact:
    name: Performa Team
servers:
  - url: /
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        count:
          type: integer
        pageCount:
          type: integer
        hasPreviousPage:
          type: boolean
        hasNextPage:
          type: boolean
...
    Teacher:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        uniqueId:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        profilePictureUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          type: object
    ProfileResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: object
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticate with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/LoginResponse'
  /auth/register-admin:
    post:
      tags: [Auth]
      summary: Register admin
      description: Register a new admin user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, roleIds]
              properties:
                username:
                  type: string
                  example: johnadmin
                email:
                  type: string
                  format: email
                  example: admin@performa.edu
                password:
                  type: string
                  example: SecurePass123!
                roleIds:
                  type: array
                  items:
                    type: string
                  example: ['clxxx_role_admin']
      responses:
        '201':
          description: Admin registered
  /auth/register-teacher:
    post:
      tags: [Auth]
      summary: Register teacher
      description: Register a new teacher with user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, teacher]
              properties:
                user:
                  type: object
                  required: [username, email, password, roleIds]
                  properties:
                    username:
                      type: string
                      example: janedoe
                    email:
                      type: string
                      format: email
                      example: jane@example.com
                    password:
                      type: string
                      example: MyPassword123!
                    roleIds:
                      type: array
                      items:
                        type: string
                      example: ['clxxx_role_teacher']
                teacher:
                  type: object
                  required: [fullName, phoneNumber]
                  properties:
                    fullName:
                      type: string
                      example: Jane Doe
                    phoneNumber:
                      type: string
                      example: '081234567890'
                    dateOfBirth:
                      type: string
                      format: date
                      example: '1995-06-15'
                    bio:
                      type: string
                      example: A passionate learner
      responses:
        '201':
          description: Teacher registered
  /auth/getMe:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ProfileResponse'

  # ── Teachers ──
  /teachers:
    get:
      tags: [Teachers]
      summary: Get all teachers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        '200':
          description: Teachers retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Teacher'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
  /teachers/{id}:
    get:
      tags: [Teachers]
      summary: Get teacher by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Teacher'
    put:
      tags: [Teachers]
      summary: Update teacher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                phoneNumber:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                bio:
                  type: string
                profilePictureUrl:
                  type: string
      responses:
        '200':
          description: Teacher updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Teacher'
    delete:
      tags: [Teachers]
      summary: Delete teacher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Teacher deleted

  # ── Contents ──
...