# types/proto/teacher-service.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.8.3
//   protoc               v7.34.0
// source: proto/teacher-service.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "teacherservice";

export enum Order {
  ASC = 0,
  DESC = 1,
  UNRECOGNIZED = -1,
}

export interface Filter {
  values: { [key: string]: string };
}

export interface Filter_ValuesEntry {
  key: string;
  value: string;
}

export interface PageMeta {
  page: number;
  pageSize: number;
  count: number;
  pageCount?: number | undefined;
  hasPreviousPage?: boolean | undefined;
  hasNextPage?: boolean | undefined;
}

export interface Role {
  id: string;
  name: string;
  permissions: string[];
}

export interface User {
  id: string;
  username: string;
  email: string;
  roles: Role[];
}

export interface Teacher {
  id: string;
  uniqueId: string;
  fullName: string;
  phoneNumber: string;
  dateOfBirth?: string | undefined;
  user?: User | undefined;
  branchId?: string | undefined;
  branchName?: string | undefined;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
}

export interface CreateTeacherRequest {
  userId: string;
  fullName: string;
  phoneNumber: string;
  branchId: string;
  branchName: string;
  dateOfBirth?: string | undefined;
}

export interface CreateTeacherResponse {
  teacher?: Teacher | undefined;
}

export interface GetTeacherByIdRequest {
  id: string;
}

export interface GetTeacherByIdResponse {
  teacher?: Teacher | undefined;
}

export interface UpdateTeacherRequest {
  id: string;
  fullName?: string | undefined;
  phoneNumber?: string | undefined;
  dateOfBirth?: string | undefined;
  branchId?: string | undefined;
  branchName?: string | undefined;
}

export interface UpdateTeacherResponse {
  teacher?: Teacher | undefined;
}

export interface DeleteTeacherRequest {
  id: string;
}

export interface DeleteTeacherResponse {
  message: string;
}

export interface GetAllTeachersRequest {
  page?: number | undefined;
  pageSize?: number | undefined;
  order?: Order | undefined;
  skip?: number | undefined;
  filter?: Filter | undefined;
  sortBy?: string | undefined;
  query?: string | undefined;
  search?: string | undefined;
}

export interface GetAllTeachersResponse {
  teachers: Teacher[];
  meta?: PageMeta | undefined;
}

export const TEACHERSERVICE_PACKAGE_NAME = "teacherservice";

/** Teacher Service Definition */

export interface TeacherServiceClient {
  /** Teacher CRUD Operations */

  createTeacher(request: CreateTeacherRequest): Observable<CreateTeacherResponse>;

  /** Retrieve a teacher by their unique ID */

  getTeacherById(request: GetTeacherByIdRequest): Observable<GetTeacherByIdResponse>;

  /** Update an existing teacher's information */

  updateTeacher(request: UpdateTeacherRequest): Observable<UpdateTeacherResponse>;

  /** Delete a teacher by their unique ID */

  deleteTeacher(request: DeleteTeacherRequest): Observable<DeleteTeacherResponse>;

  /** Retrieve all teachers with optional pagination, filtering, and sorting */

  getAllTeachers(request: GetAllTeachersRequest): Observable<GetAllTeachersResponse>;
}

/** Teacher Service Definition */

export interface TeacherServiceController {
  /** Teacher CRUD Operations */

  createTeacher(
    request: CreateTeacherRequest,
  ): Promise<CreateTeacherResponse> | Observable<CreateTeacherResponse> | CreateTeacherResponse;

  /** Retrieve a teacher by their unique ID */

  getTeacherById(
    request: GetTeacherByIdRequest,
  ): Promise<GetTeacherByIdResponse> | Observable<GetTeacherByIdResponse> | GetTeacherByIdResponse;

  /** Update an existing teacher's information */

  updateTeacher(
    request: UpdateTeacherRequest,
  ): Promise<UpdateTeacherResponse> | Observable<UpdateTeacherResponse> | UpdateTeacherResponse;

  /** Delete a teacher by their unique ID */

  deleteTeacher(
    request: DeleteTeacherRequest,
  ): Promise<DeleteTeacherResponse> | Observable<DeleteTeacherResponse> | DeleteTeacherResponse;

  /** Retrieve all teachers with optional pagination, filtering, and sorting */

  getAllTeachers(
    request: GetAllTeachersRequest,
  ): Promise<GetAllTeachersResponse> | Observable<GetAllTeachersResponse> | GetAllTeachersResponse;
}

export function TeacherServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = [
      "createTeacher",
      "getTeacherById",
      "updateTeacher",
      "deleteTeacher",
      "getAllTeachers",
    ];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("TeacherService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("TeacherService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const TEACHER_SERVICE_NAME = "TeacherService";